Refactor opCodes
Some operations require a different number of paramaters after the 1byte opCode.
In the current implementation, this cannot be done, because we would need a
function that takes an arbitrary number of parameters.

For this reason, I refactored the code so that the opCode now takes the memory as
input, computes the address of the current operation and reads the opCode directly
from the memory. Based on the opCode, if extra arguments are required, those are
also read from the memory.
At the end, the IP register is updated based on how much memory has been read.